pyproject.toml
[tool.poetry]
name = "dbt-artifacts-ext"
version = "0.0.1"
description = "`dbt-artifacts-ext` is a Meltano utility extension."
authors = ["Reuben Frankel"]
license = "Apache 2.0"
packages = [
    { include = "dbt_artifacts_ext" }
    # uncomment this if you want to include static assets
    # { include = "files_dbt-artifacts-ext.*" }
]
# If you need to static assets with your extension, you can add them here.
# Under a directory path of "files_dbt-artifacts-ext/"
#include = [
#    "files_dbt-artifacts-ext/somefile.thing"
#]

[tool.poetry.dependencies]
python = "<3.12,>=3.7"
PyYAML = "^6.0.0"
click = "^8.1.3"
typer = "^0.6.1"
"meltano.edk"= {git = "https://github.com/meltano/edk.git", rev="main"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dbt-artifacts_extension = 'dbt_artifacts_ext.main:app'
